// src/types/travel.ts
export interface Landmark {
  id: string;
  city_code: string;
  country_code: string;
  image: string;
  images: string[];
  landmark_code: string;
  latitude: number;
  longitude: number;
  type: string;
  l_ar_description: string;
  l_ar_name: string;
  l_ar_wikipedia: string;
  l_en_address: string;
  l_en_description: string;
  l_en_name: string;
  l_en_types: string;
  l_en_wikipedia: string;
  l_es_description: string;
  l_es_name: string;
  l_es_wikipedia: string;
  l_fr_description: string;
  l_fr_name: string;
  l_fr_wikipedia: string;
  l_zh_description: string;
  l_zh_name: string;
  l_zh_wikipedia: string;
}